Automatically backed up by Learn

+++ lib/chance_of_rain.rb
@@ -1,6 +1,6 @@
 puts "What is the chance of rain? Enter percent chance as a decimal."
 
-chance_of_rain = gets.strip.to_i
+chance_of_rain = gets.strip.to_f
 if chance_of_rain <= 0.25
   puts "Pack a sun shelter!"
 elsif (chance_of_rain > 0.25 && chance_of_rain < 0.75)
